+++ src/utils/jsongen.js
@@ -1,0 +1,17 @@
+import axios from "axios";
+import crypto from 'crypto';
+import UserAgent from 'fake-useragent';
+
+export const jsongen = async (url) => {
+  try {
+    const headers = {
+      'X-Signature-Version': 'web2',
+      'X-Signature': crypto.randomBytes(32).toString('hex'),
+      'User-Agent': new UserAgent().random,
+    };
+    const res = await axios.get(url, { headers });
+    return res.data;
+  } catch (error) {
+    throw new Error(`Error fetching data: ${error.message}`);
+  }
+};
